# openapi.yaml
openapi: 3.0.3
info:
  title: SNS Backend API
  version: 1.0.0
  description: |
    OpenAPI schema generated from the current Express implementation in index.js.
servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Auth
    description: Authentication and user registration
  - name: Posts
    description: Post CRUD operations
  - name: Users
    description: User profile operations

paths:
  /login:
    post:
      tags: [Auth]
      summary: Login
      description: Authenticates a user and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
                    description: JWT access token
        '400':
          description: Validation error or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /register:
    post:
      tags: [Auth]
      summary: Register user
      description: Creates a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      description: Returns a user's public profile by user id.
      parameters:
        - in: path
          name: id
          required: true
          description: User id
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/{userId}/follow-summary:
    get:
      tags: [Users]
      summary: Get follow summary for a user
      description: Returns follow counts and current user's follow state for the profile user.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          description: Profile user id
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowSummary'
        '400':
          description: Invalid userId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/{userId}/follow:
    post:
      tags: [Users]
      summary: Follow a user
      description: Follows the target user. Idempotent and requires bearer authentication.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          description: User id to follow
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Follow action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowActionResponse'
        '400':
          description: Invalid userId or self-follow is prohibited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Users]
      summary: Unfollow a user
      description: Unfollows the target user. Idempotent and requires bearer authentication.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          description: User id to unfollow
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Unfollow action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowActionResponse'
        '400':
          description: Invalid userId or self-unfollow is prohibited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /posts:
    get:
      tags: [Posts]
      summary: List posts
      description: Returns all posts ordered by id descending.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Posts]
      summary: Create post
      description: Creates a post. Requires Bearer token.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpsertRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                type: object
                required: [message, postId]
                properties:
                  message:
                    type: string
                  postId:
                    type: integer
                    minimum: 1
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /posts/{id}:
    put:
      tags: [Posts]
      summary: Update post
      description: Updates a post by id. Requires Bearer token.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Post id
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpsertRequest'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid id or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Posts]
      summary: Delete post
      description: Deletes a post by id. Requires Bearer token.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Post id
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Post deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server-side error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1

    Post:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        title:
          type: string
        content:
          type: string
      additionalProperties: true

    User:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
          minimum: 1
        username:
          type: string

    PostUpsertRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    FollowSummary:
      type: object
      required: [userId, followersCount, followingCount, isFollowing]
      properties:
        userId:
          type: integer
          minimum: 1
        followersCount:
          type: integer
          minimum: 0
        followingCount:
          type: integer
          minimum: 0
        isFollowing:
          type: boolean

    FollowActionResponse:
      type: object
      required: [ok, isFollowing, followersCount]
      properties:
        ok:
          type: boolean
        isFollowing:
          type: boolean
        followersCount:
          type: integer
          minimum: 0

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string